ai-service/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
openai==1.3.7
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# Code Analysis
ast==0.0.2
black==23.11.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Vector Database and Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Database and Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
python-dotenv==1.0.0
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
pre-commit==3.6.0

# Security
cryptography==41.0.8
passlib==1.7.4
python-jose[cryptography]==3.3.0

# Async and Concurrency
asyncio-mqtt==0.16.1
celery==5.3.4

# Code Quality
bandit==1.7.5
safety==2.3.5 